Cleanup and Commenting added
